pyproject.toml
[tool.poetry]
name = "openstack_swift_api"
version = "0.1.0"
description = ""
authors = ["Will Siddall <will.siddall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-swiftclient = "^3.12.0"
Flask = "^2.0.2"
python-dotenv = "^0.19.1"
python-keystoneclient = "^4.3.0"
psycopg2-binary = "^2.9.1"
alembic = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
